extracted inner panels to compile on IC-2022.1
