+++ src/main/java/org/jboss/tools/intellij/openshift/ui/helm/ChartPanel.java
@@ -1,0 +1,5 @@
+package org.jboss.tools.intellij.openshift.ui.helm;
+
+interface ChartPanel {
+    void setChart(ChartVersions chart);
+  }
